requirements.txt
boto3
strands-agents
strands-agents-tools
strands-agents[ollama]
PyPDF2
streamlit